add view and delete

+++ app/Controllers/Students.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 
 namespace App\Controllers;
 
@@ -7,41 +7,58 @@ use App\Models\StudentModel;
 class Students extends BaseController
 {
     public function index()
-    { 
-       // instance of StudentModel
-       $studentModel = new StudentModel();
-       
-       //fetch all students 
-       $data['students'] = $studentModel->findAll();
+    {
+        // instance of StudentModel
+        $studentModel = new StudentModel();
 
-     return view('student_registration', $data);
+        //fetch all students 
+        $data['students'] = $studentModel->findAll();
+
+        return view('student_registration', $data);
     }
 
     public function create()
     {
-       $studentModel = new StudentModel();
+        $studentModel = new StudentModel();
 
-       $data = [
+        $data = [
             'name' => $this->request->getPost('name'),
-            'email' => $this->request->getPost('email'),  
-            'address' => $this->request->getPost('address'), 
+            'email' => $this->request->getPost('email'),
+            'address' => $this->request->getPost('address'),
             'phone' => $this->request->getPost('phone'),
             'father_name' => $this->request->getPost('father_name'),
             'mother_name' => $this->request->getPost('mother_name'),
-       ];
-       $studentModel->save($data);
+        ];
+        $studentModel->save($data);
 
-       return redirect()->to('/students');
+        return redirect()->to('/students');
     }
 
     public function delete($id = null)
     {
-        //instance of StudentModel
+
+
+        if ($id!= 2) {
+            return "Access Denied: You do not have permission to delete students.";
+        }
         $studentModel = new StudentModel();
 
-        if($id !== null){
-            $studentModel->delete($id);  
+        if ($id !== null) {
+            $studentModel->delete($id);
         }
         return redirect()->to('/students');
+    }
+
+    public function view($id = null)
+    {
+        $studentModel = new StudentModel();
+
+        $data['student'] = $studentModel->find($id);
+
+        if (empty($data['student'])) {
+            throw new \CodeIgniter\Exceptions\PageNotFoundException('Cannot find the student item: ' . $id);
+        }
+
+        return view('student_view', $data);
     }
 }
